refactor(web): split workspace-table into per-concern files
Move the workspace-table monolith into a directory with focused
modules: headers, rows, end fillers, and shared internals. Entry stays
at workspace-table/index.tsx; resolves without consumer changes.
